requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
streamlit>=1.38
python-dotenv>=1.0
